Merge Tencent-TDS/KuiklyUI main (19e4a721) into the fork. Every fork
customization was checked against upstream to drop the ones upstream now
covers, so the divergence is only what upstream still lacks.

Conflicts resolved:
- PagerStateExtensions.kt: take upstream, fork version was byte-identical
- ScrollerView.kt: take upstream, fork's pagingEnabled() duplicated the
  pre-existing pagingEnable()
- SubcomposeLayout.kt: take upstream verbatim, see H5 pager below
- KRView.kt (android): combine fork ripple support with upstream text selection

Dropped as superseded by upstream:
- pager boundary check and HorizontalPager initialPage sync, upstream rewrote
  the pager with a more complete fix
- PlatformUtils.isWeb/isMiniApp, PagerData.isWeb/isMiniApp predate the fork
  point by five months and the helpers had zero callers

Reworked H5 compose pager snapping:
- setting pagingEnable from compose commonMain reached all five render targets
  and was never re-applied on node reuse, leaving stale paging state that froze
  reused lists
- route upstream's existing isComposePager hint through the web render layer
  instead: IListElement.setComposePager defaults to unhandled, H5ListView maps
  it to paging, mini-program behaviour unchanged

Kept, upstream has no equivalent:
- native ripple for Android, iOS and Web
- KSP JS/H5 entry generation and metadata target skip
- KRVideoView canPlayType handling
- H5ListView scroll event dedup and H5ListPagingHelper fixes
- Gradle 8.7 and JVM 21 build configuration

+++ core/src/commonMain/kotlin/com/tencent/kuikly/core/base/Color.kt
@@ -85,7 +85,7 @@ class Color {
 
     override fun toString(): String {
         return colorString.ifEmpty {
-            _hexColor.toString()
+            _hexColor.toUInt().toString()
         }
     }
